service/Services/entities/src/main/java/org/infy/entities/artifact/ArtifactDetails.java
/***********************************************************************************************
*
* Copyright 2018 Infosys Ltd.
* Use of this source code is governed by MIT license that can be found in the LICENSE file or at
* https://opensource.org/licenses/MIT.
*
***********************************************************************************************/

package org.infy.entities.artifact;

import com.google.gson.annotations.Expose;
import com.google.gson.annotations.SerializedName;

/**
 * Entity to store artifact related details
 * 
 * @author Infosys
 *
 */
public class ArtifactDetails {

	@SerializedName("status")
	@Expose
	private String status;
	@SerializedName("remark")
	@Expose
	private String remark;
	@SerializedName("environment")
	@Expose
	private String environment;
	@SerializedName("owner")
	@Expose
	private String owner;
	@SerializedName("actionTime")
	@Expose
	private String actionTime;

	public String getStatus() {
		return status;
	}

	public void setStatus(String status) {
		this.status = status;
	}

	public String getRemark() {
		return remark;
	}

	public void setRemark(String remark) {
		this.remark = remark;
	}

	public String getEnvironment() {
		return environment;
	}

	public void setEnvironment(String environment) {
		this.environment = environment;
	}

	public String getOwner() {
		return owner;
	}

	public void setOwner(String owner) {
		this.owner = owner;
	}

	public String getActionTime() {
		return actionTime;
	}

	public void setActionTime(String actionTime) {
		this.actionTime = actionTime;
	}

}
